extract common test harness code
